+++ HISTORY.rst
@@ -5,6 +5,8 @@ History
 1.7.0 (2020-04-??)
 ------------------
 * Added some functions to read table data from cube files.
+* Added special pixel values via specialpixels.py (implemented
+  as namedtuples instead of as dictionaries, as in pysis).
 * Enabled Python 3.7 and 3.8 tests for tox.
 
 1.6.0 (2020-04-12)